+++ src/main/java/com/telepathicgrunt/repurposedstructures/world/structures/OutpostNetherStructure.java
@@ -21,7 +21,7 @@ import net.minecraft.world.gen.feature.template.TemplateManager;
 import java.util.List;
 
 
-public class OutpostNetherStructure extends AbstractBaseStructure {
+public class OutpostNetherStructure extends AbstractBaseStructure<NoFeatureConfig> {
     private final ResourceLocation START_POOL;
     private static final List<MobSpawnInfo.Spawners> MONSTER_SPAWNS = Lists.newArrayList(new MobSpawnInfo.Spawners(EntityType.PIGLIN, 10, 1, 1));